feat: keep Deployment environments updated (#218)
Update the Deployment associated with a ControlPlane or DataPlane if its
current actual EnvVar set or EnvFrom differs from the set generated by the
reconciler.

Add Service information to DataPlane status.

Have the ControlPlane reconciler watch for changes to the controlled
DataPlane. If the DataPlane status contains a new Service, the generated
EnvVars will differ and the reconciler will update the ControlPlane with
the new publish service.

Avoid a race condition by wrapping the CFSSL logger override in a
sync.Once.

Retrieve containers by their name rather than relying on indices. Fail
if we cannot find a container with that name.